requirements.txt
albucore==0.0.13
albumentations==1.4.8
annotated-types==0.7.0
causal-conv1d==1.0.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
cmake==3.29.2
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
exceptiongroup==1.2.1
filelock==3.13.4
fonttools==4.51.0
fsspec==2024.3.1
h5py==3.11.0
huggingface-hub==0.22.2
idna==3.7
imageio==2.34.1
imgaug==0.4.0
importlib_resources==6.4.0
iniconfig==2.0.0
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
lit==18.1.4
mamba-ssm==1.0.1
matplotlib==3.7.5
MedPy==0.5.1
networkx==3.1
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
protobuf==5.26.1
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
shapely==2.0.4
SimpleITK==2.3.1
six==1.16.0
submitit==1.5.1
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2023.7.10
timm==0.4.12
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.0
torchaudio==0.13.0
torchvision==0.14.0
tqdm==4.66.2
transformers==4.40.1
triton==2.0.0
typing_extensions==4.11.0
urllib3==2.2.1
yacs==0.1.8
zipp==3.18.1
